+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/manager/EthProtocolManager.java
@@ -175,6 +175,7 @@ public class EthProtocolManager implements ProtocolManager, MinedBlockObserver {
 
     capabilities.add(EthProtocol.ETH68);
     capabilities.add(EthProtocol.ETH69);
+    capabilities.add(EthProtocol.ETH70);
     capabilities.removeIf(cap -> cap.getVersion() > ethProtocolConfiguration.getMaxEthCapability());
     capabilities.removeIf(cap -> cap.getVersion() < ethProtocolConfiguration.getMinEthCapability());